requirements.txt
python-nmap
python-whois
pyfiglet
colorama
argparse
sys
